# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个基于底层库构建的跨平台音乐搜索与解析统一 API 服务。",
        "title": "Go Music API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/music/cover": {
            "get": {
                "description": "发送带伪造标头的请求拉取远端封面大图，避开网易云、QQ 音乐的图片防盗链 403 问题。",
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "代理请求并下载封面图",
                "parameters": [
                    {
                        "type": "string",
                        "description": "封面图原始 URL (需经过 urlencode)",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "歌曲名(用于生成下载文件名)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "歌手名(用于生成下载文件名)",
                        "name": "artist",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "原封不动的图片流",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/music/inspect": {
            "get": {
                "description": "快速探测音频直链的可访问性，并根据 `Content-Range` 推算文件大小及大概码率。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "探测音频大小与码率",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1815969416\"",
                        "description": "音乐 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"netease\"",
                        "description": "音乐来源平台",
                        "name": "source",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"255\"",
                        "description": "音乐时长(秒)，若提供则可更精确预估码率(kbps)",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "包含有效状态、真实URL、文件大小和码率等探测信息",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/music/lyric": {
            "get": {
                "description": "抓取对应歌曲的带有时间轴的完整 LRC 歌词文本，采用标准 JSON 格式返回。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "获取 JSON 格式歌词",
                "parameters": [
                    {
                        "type": "string",
                        "description": "音乐 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "包含 lyric 字符串属性的数据对象",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "对应平台未实现歌词抓取",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/music/lyric/file": {
            "get": {
                "description": "作为附件直接下载 `.lrc` 后缀的歌词文件到本地。",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "下载 LRC 歌词文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "音乐 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台",
                        "name": "source",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "音乐名称 (用于生成保存文件名)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "歌手名称 (用于生成保存文件名)",
                        "name": "artist",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "纯文本文件流",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/music/search": {
            "get": {
                "description": "兼容多源并发搜索以及链接智能解析，自动返回单曲或歌单数组。支持直接输入关键词或粘贴音乐平台的分享链接。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "综合搜索与链接解析",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"七里香 周杰伦\"",
                        "description": "关键词或音乐分享链接",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "song",
                            "playlist"
                        ],
                        "type": "string",
                        "default": "\"song\"",
                        "description": "搜索类型: song (单曲) 或 playlist (歌单)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "指定的音源数组(留空则默认全平台)。例: netease, qq",
                        "name": "sources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回解析的数据，包含歌曲/歌单列表",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "不支持的链接解析",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "解析过程出现错误",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/music/url": {
            "get": {
                "description": "获取解析到的原始音频播放链接。注：客户端直接使用裸链接可能受到平台的防盗链拦截限制（推荐使用 `/stream` 代理接口）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Music"
                ],
                "summary": "获取音频裸直链",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"1815969416\"",
                        "description": "音乐 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"netease\"",
                        "description": "平台源",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "直接返回带有 url 的数据实体",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "源不支持",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "链接抓取失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}